refactor(tests): split client modules and files to satisfy size and coverage gates
